# tsp-output/@typespec/openapi3/openapi.yaml
openapi: 3.0.0
info:
  title: Booking API
  version: 0.0.0
...
paths:
...
  /event-types:
    get:
      operationId: EventTypeOperations_listEventTypes
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
              example:
                - id: "evt-15min"
                  title: "Встреча 15 минут"
                  description: "Короткий тип события для быстрого слота."
                  durationMinutes: 15
                - id: "evt-30min"
                  title: "Встреча 30 минут"
                  description: "Базовый тип события для бронирования."
                  durationMinutes: 30
    post:
      operationId: EventTypeOperations_createEventType
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventType'
  /event-types/{eventTypeId}:
    get:
      operationId: EventTypeItemOperations_getEventType
      parameters:
        - name: eventTypeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
...
components:
  schemas:
    BadRequestError:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          enum:
            - BAD_REQUEST
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
...
    ConflictError:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          enum:
            - CONFLICT
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: string
    EventType:
      type: object
      required:
        - id
        - title
        - description
        - durationMinutes
      properties:
        id:
          type: string
          example: "evt-15min"
        title:
          type: string
          example: "Встреча 15 минут"
        description:
          type: string
          example: "Короткий тип события для быстрого слота."
        durationMinutes:
          type: integer
          format: int32
          example: 15
    NotFoundError:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          enum:
            - NOT_FOUND
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
...
servers:
  - url: https://api.example.com
    description: Production server
    variables: {}